xhosa translation of problem solving video.
This is a really ugly hack to get it in there. The reason for this is to minimize the amount of code that has to get replaced on the client machines. Since deploying this change will involve having the coordinator drop one video file into the right directory and replace one template file, that keeps things much simpler than a proper implementation, which would require changing the watchvideo_game model to allow multiple videos per game, pushing out a change to the application data, etc. Simpler means less likely to break on the other side of the world. In this case, I think that's a worthwhile tradeoff.