speed up manage page. PMT #101570
The sheer number of backups that have accumulated in the system are making
the manage page slow to load, causing nginx to time out and 502 sometimes.

There's really no reason admins need direct access to every single backup
file right there. Limiting it to 20 speeds things up and should be
fine. If anyone really needs older backups, we still have them in the
database, so we can get them.