go straight to practice mode
per #110536, for the near future, we want to put masivukeni into "demo
mode" only. So when someone visits the site, they can just go straight
into a practice session, so they can experience it and get a feel for
it and we can easily demo it. But since we're not deploying to laptops,
the "real" mode and everything around managing participants should be
hidden.

This is the first step on that.

* link on main page jumps straight to the practice session
* breadcrumbs and headers hide the usual links to counselor management
  pages and such.

as discussed in #110537, there is still work to be done to properly
support concurrent practice sessions in the online version. That work
will be done seperately.

Similarly, this version still requires the user to login via CAS (or we
can setup a 'demo/demo' type account pretty easily) even to run the
practice session. We also probably want to remove that restriction. The
code assumes a logged in counselor throughout though, so it will take
some work to allow anonymous usage of the practice session. Again, that
will be done seperately.