stub for intervention report
PMT #110459

There's going to be some design back and forth as we figure out exactly
what will go into the report and how it should be formatted, etc.

This just gets a stub in place showing how many minutes were spent on
each session/activity to get the ball rolling.