PMT #112840: Questions and Concerns field not saving
This functionality was not totally finished when this site was moved to
a 'stateless' site. The main issue was that a CounselorNote object
required a Counselor. In the initial run of this site, a counselor would
be logged in to walk the participant through the intervention. Because
this is now stateless, there is no logged in Counselor.

This commit fixes the issue by making the counselor field nullable on
the ConselorNote object.